Add Json Benchmark
