+++ concepts/serialiser/CMakeLists.txt
@@ -1,5 +1,8 @@
-add_executable(IoClassSerialiserBenchmark IoClassSerialiserBenchmark.cpp IoClassSerialiserBenchmark.hpp)
-target_link_libraries(IoClassSerialiserBenchmark PRIVATE serialiser project_options project_warnings)
+add_executable(IoSerialiserYaSBenchmark IoSerialiserYaSBenchmark.cpp IoSerialiserBenchmark.hpp)
+target_link_libraries(IoSerialiserYaSBenchmark PRIVATE serialiser project_options project_warnings)
+
+add_executable(IoSerialiserJsonBenchmark IoSerialiserJsonBenchmark.cpp IoSerialiserBenchmark.hpp)
+target_link_libraries(IoSerialiserJsonBenchmark PRIVATE serialiser project_options project_warnings)
 
 add_executable(SimpleReflectionExample SimpleReflectionExample.cpp)
 target_link_libraries(SimpleReflectionExample PRIVATE serialiser project_options project_warnings)
